[BE] Feat: MainPage tale 목록, 상세구현 (S08P22A601-23)

+++ backend/doeng/src/main/java/com/ssafy/doeng/data/dto/tale/request/RequestTaleDetailDto.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 public class RequestTaleDetailDto {
 
     private long memberId;
-    private int taleId;
+    private long taleId;
 }
